Player Movement and Animations Course Content Finished
Wrapped up player movement for now, animations are all done with minimal foot sliding, blended walk and run animations and controller input. There's a persistent issue of the idle animation speeding way up when going really slowly, may be something to do with how Unreal's blending the idle and walk animations but I'm putting it to one side for now.