AI Moving and Shooting (Suicide Bug Active)
Implemented a number of behavior tree nodes and services in C++ to manipulate the AI's behavior tree. AI now runs after the character while within line of sight, shoots the character when within an acceptable range, and retreats back to their spawn location after a short waiting period. The AI character has a bug where it only hit's itself when it shoots, and keeps their behaviour after death as well.