Gun Aim, Jump Animations & State Machines
Added some animation for aiming the gun up or down, tweaking AI functionality to allow them to benefit as well. Additionally animations for jumping and landing were added as a state machine, and the previously implemented death animation was also bundled into a state machine containing the rest of the animations so far.